requirements/base.txt
django-authtools==1.6.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django<2.3
easy-thumbnails==2.5
pillow==5.3.0
pytz==2018.7
django-solo